# Make ImmutableRecord the only type of record