--- a/bindings/go/rs_src/types.rs
+++ b/bindings/go/rs_src/types.rs
@@ -142,6 +142,25 @@ impl LimboValue {
         Box::into_raw(Box::new(self)) as *const c_void
     }
 
+    pub fn from_owned_value(value: &limbo_core::OwnedValue) -> Self {
+        match value {
+            limbo_core::OwnedValue::Integer(i) => {
+                LimboValue::new(ValueType::Integer, ValueUnion::from_int(*i))
+            }
+            limbo_core::OwnedValue::Float(r) => {
+                LimboValue::new(ValueType::Real, ValueUnion::from_real(*r))
+            }
+            limbo_core::OwnedValue::Text(s) => {
+                LimboValue::new(ValueType::Text, ValueUnion::from_str(s.as_str()))
+            }
+            limbo_core::OwnedValue::Blob(b) => {
+                LimboValue::new(ValueType::Blob, ValueUnion::from_bytes(b.as_slice()))
+            }
+            limbo_core::OwnedValue::Null => {
+                LimboValue::new(ValueType::Null, ValueUnion::from_null())
+            }
+        }
+    }
     pub fn from_value(value: &limbo_core::RefValue) -> Self {
         match value {
             limbo_core::RefValue::Integer(i) => {