--- a/bindings/javascript/src/lib.rs
+++ b/bindings/javascript/src/lib.rs
@@ -92,14 +92,14 @@ impl Statement {
     }
 }
 
-fn to_js_value(env: &napi::Env, value: &limbo_core::OwnedValue) -> JsUnknown {
+fn to_js_value(env: &napi::Env, value: &limbo_core::RefValue) -> JsUnknown {
     match value {
-        limbo_core::OwnedValue::Null => env.get_null().unwrap().into_unknown(),
-        limbo_core::OwnedValue::Integer(i) => env.create_int64(*i).unwrap().into_unknown(),
-        limbo_core::OwnedValue::Float(f) => env.create_double(*f).unwrap().into_unknown(),
-        limbo_core::OwnedValue::Text(s) => env.create_string(s.as_str()).unwrap().into_unknown(),
-        limbo_core::OwnedValue::Blob(b) => {
-            env.create_buffer_copy(b.as_ref()).unwrap().into_unknown()
+        limbo_core::RefValue::Null => env.get_null().unwrap().into_unknown(),
+        limbo_core::RefValue::Integer(i) => env.create_int64(*i).unwrap().into_unknown(),
+        limbo_core::RefValue::Float(f) => env.create_double(*f).unwrap().into_unknown(),
+        limbo_core::RefValue::Text(s) => env.create_string(s.as_str()).unwrap().into_unknown(),
+        limbo_core::RefValue::Blob(b) => {
+            env.create_buffer_copy(b.to_slice()).unwrap().into_unknown()
         }
     }
 }