--- a/cli/app.rs
+++ b/cli/app.rs
@@ -5,7 +5,7 @@ use crate::{
     opcodes_dictionary::OPCODE_DESCRIPTIONS,
 };
 use comfy_table::{Attribute, Cell, CellAlignment, Color, ContentArrangement, Row, Table};
-use limbo_core::{Database, LimboError, RefValue, Statement, StepResult};
+use limbo_core::{Database, LimboError, OwnedValue, RefValue, Statement, StepResult};
 
 use clap::{Parser, ValueEnum};
 use rustyline::{history::DefaultHistory, Editor};
@@ -704,18 +704,18 @@ impl<'a> Limbo<'a> {
                     match rows.step() {
                         Ok(StepResult::Row) => {
                             let row = rows.row().unwrap();
-                            for (i, value) in row.get_values().iter().enumerate() {
+                            for (i, value) in row.get_values().enumerate() {
                                 if i > 0 {
                                     let _ = self.writer.write(b"|");
                                 }
                                 let _ = self.writer.write(
                                     match value {
-                                        RefValue::Null => self.opts.null_value.clone(),
-                                        RefValue::Integer(i) => format!("{}", i),
-                                        RefValue::Float(f) => format!("{:?}", f),
-                                        RefValue::Text(s) => s.to_string(),
-                                        RefValue::Blob(b) => {
-                                            format!("{}", String::from_utf8_lossy(b.to_slice()))
+                                        OwnedValue::Null => self.opts.null_value.clone(),
+                                        OwnedValue::Integer(i) => format!("{}", i),
+                                        OwnedValue::Float(f) => format!("{:?}", f),
+                                        OwnedValue::Text(s) => s.to_string(),
+                                        OwnedValue::Blob(b) => {
+                                            format!("{}", String::from_utf8_lossy(b.as_slice()))
                                         }
                                     }
                                     .as_bytes(),
@@ -767,18 +767,20 @@ impl<'a> Limbo<'a> {
                                 let record = rows.row().unwrap();
                                 let mut row = Row::new();
                                 row.max_height(1);
-                                for (idx, value) in record.get_values().iter().enumerate() {
+                                for (idx, value) in record.get_values().enumerate() {
                                     let (content, alignment) = match value {
-                                        RefValue::Null => {
+                                        OwnedValue::Null => {
                                             (self.opts.null_value.clone(), CellAlignment::Left)
                                         }
-                                        RefValue::Integer(i) => {
+                                        OwnedValue::Integer(i) => {
                                             (i.to_string(), CellAlignment::Right)
                                         }
-                                        RefValue::Float(f) => (f.to_string(), CellAlignment::Right),
-                                        RefValue::Text(s) => (s.to_string(), CellAlignment::Left),
-                                        RefValue::Blob(b) => (
-                                            String::from_utf8_lossy(b.to_slice()).to_string(),
+                                        OwnedValue::Float(f) => {
+                                            (f.to_string(), CellAlignment::Right)
+                                        }
+                                        OwnedValue::Text(s) => (s.to_string(), CellAlignment::Left),
+                                        OwnedValue::Blob(b) => (
+                                            String::from_utf8_lossy(b.as_slice()).to_string(),
                                             CellAlignment::Left,
                                         ),
                                     };
@@ -846,7 +848,7 @@ impl<'a> Limbo<'a> {
                     match rows.step()? {
                         StepResult::Row => {
                             let row = rows.row().unwrap();
-                            if let Some(RefValue::Text(schema)) = row.get_values().first() {
+                            if let Ok(OwnedValue::Text(schema)) = row.get::<&OwnedValue>(0) {
                                 let _ = self.write_fmt(format_args!("{};", schema.as_str()));
                                 found = true;
                             }
@@ -904,7 +906,7 @@ impl<'a> Limbo<'a> {
                     match rows.step()? {
                         StepResult::Row => {
                             let row = rows.row().unwrap();
-                            if let Some(RefValue::Text(table)) = row.get_values().first() {
+                            if let Ok(OwnedValue::Text(table)) = row.get::<&OwnedValue>(0) {
                                 tables.push_str(table.as_str());
                                 tables.push(' ');
                             }