--- a/core/lib.rs
+++ b/core/lib.rs
@@ -62,6 +62,7 @@ use storage::{
 };
 use translate::select::prepare_select_plan;
 pub use types::OwnedValue;
+pub use types::RefValue;
 use util::{columns_from_create_table_body, parse_schema_rows};
 use vdbe::{builder::QueryMode, VTabOpaqueCursor};
 
@@ -603,7 +604,7 @@ impl Statement {
     }
 }
 
-pub type Row = types::Record;
+pub type Row = types::ImmutableRecord;
 
 pub type StepResult = vdbe::StepResult;