This pr is huge again but I will try to introduce each improvement one by one.

## Overview
### Remove Rc for Text and Blob.

In general copying is bad, that's why we hid it with `Rc`s. With the introduction of `ImmutableRecord` we make it less relevant because now we will copy only once anyways, no other place should copy it so we can avoid using `Rc`. If we we were to copy it it most likely means where are doing something wrong. 

### Reuse `Text` and `Blob` OwnedValues.

Most of the queries spend time overwriting the same register over and over. What about we don't allocate new `OwnedValue` and we just simply reuse the `OwnedValue` and extend the internal buffer. That's what I did and it worked quite nicely.

### Make `Register::Record` be `ImmutableRecord`

`ImmutableRecord` basically means "serialized record", that's why all the data is contained in a single payload buffer. There is a list of values to reference that payload to reduce time complexity of search -- there is an argument to make a record without this vec to reduce memory footprint.  This improvement I don't think it had a direct impact on performance but it is a simpler way to lay the memory without any complicated reference counted pointers, and instead we use a contiguous piece of memory.

### Make `ImmutableRecord` reusable in `BTreeCursor`.

`BTreeCursor` allocated and deallocated records when it needed a new one. This is obviously a big waste because we could be reusing the internal buffer to avoid allocations. `ImmutableRecord` proved to be useful here because now, we will only store a single `ImmutableRecord` in the cursor that we will never deallocate -- we will just reallocate when needed and replace the current one with the next one on demand.

## Return `Row` as a reference of Registers.

A `ResultRow` bytecode takes care of gathering all the columns of a row and returning them to the user. Previously we could create a new `Record` struct with all the cloned values which proved to be wasteful. SQLite is smart about this so we must be as well. Basically a row now is a wrapper for `struct Row { *const Register, count: usize }`, and we basically include some QOL methods to avoid using pointers directly. 

I know pointers are unsafe. That's why this row will be invalidate on the next step of the VM and this row should be not used outside there. 

### Inlining go brrr
`read_varint` and `read_value` are called in a tight loop making it easy to see overhead of the call stack. That's why I sprinkled some `#[inline(always)]` and saw something like a 15% speed boost.

## read_record with custom `SmallVec<T>`

We tend to overuse vectors for everything, this is quite bad because it requires heap memory allocations. We can avoid this with a simple `SmallVec` that simply fallsback to a vec with more complex scenarios. 


## Benchmarks!

```
### before
fun/limbo » cargo bench -- limbo_execute 2>&1 | grep -B 1 "time: " | tee out.log
Execute `SELECT 1`/limbo_execute_select_1
                        time:   [43.958 ns 44.056 ns 44.154 ns]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/1
                        time:   [407.82 ns 408.57 ns 409.41 ns]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/10
                        time:   [2.7335 µs 2.7386 µs 2.7443 µs]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/50
                        time:   [13.451 µs 13.485 µs 13.520 µs]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/100
                        time:   [26.967 µs 27.077 µs 27.201 µs]after:
```
### after
```
fun/limbo (more-register) » cargo bench -- limbo_execute 2>&1 | grep -B 1 "time: " | tee out.log                                                                                                                                                                                                                                                        130 ↵
Execute `SELECT 1`/limbo_execute_select_1
                        time:   [33.317 ns 33.458 ns 33.645 ns]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/1
                        time:   [336.62 ns 337.21 ns 337.81 ns]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/10
                        time:   [1.6860 µs 1.6911 µs 1.6960 µs]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/50
                        time:   [7.8916 µs 7.9160 µs 7.9468 µs]
--
Execute `SELECT * FROM users LIMIT ?`/limbo_execute_select_rows/100
                        time:   [15.682 µs 15.717 µs 15.754 µs]
```